requirements.txt
numpy==1.26.1
pygame-ce==2.3.2
pygame-gui==0.6.9
python-i18n==0.3.9
